+++ Dockerfile
@@ -1,6 +1,6 @@
 FROM python:3.7
 
-RUN pip install pytest
+RUN pip install pytest pytest-cases
 
 
 WORKDIR /usr/src/labelbox
