+++ tests/integration/bulk_import/test_bulk_import_request.py
@@ -1,6 +1,4 @@
 import uuid
-from attr import validate
-
 import ndjson
 import pytest
 import requests
