+++ tests/integration/bulk_import/test_ndjson_validation.py
@@ -1,11 +1,13 @@
-from labelbox.exceptions import NDJsonError, UuidError
+import pytest
+import ndjson
+from pytest_cases import pytest_parametrize_plus, fixture_ref
+
+from labelbox.exceptions import ValidationError
 from labelbox.schema.bulk_import_request import (NDCheckList, NDClassification,
                                                  NDMask, NDPolygon, NDPolyline,
                                                  NDRadio, NDRectangle, NDText,
                                                  NDTextEntity, NDTool,
                                                  _validate_ndjson)
-import pytest
-import ndjson
 
 
 def test_classification_construction(checklist_inference, text_inference):
@@ -21,24 +23,14 @@ def test_subclassification_construction(rectangle_inference):
     assert isinstance(tool.classifications[0], NDRadio)
 
 
-def test_tool_construction(polygon_inference, rectangle_inference,
-                           line_inference, entity_inference,
-                           segmentation_inference):
-
-    polygon = NDTool.build(polygon_inference)
-    assert isinstance(polygon, NDPolygon)
-
-    rectangle = NDTool.build(rectangle_inference)
-    assert isinstance(rectangle, NDRectangle)
-
-    line = NDTool.build(line_inference)
-    assert isinstance(line, NDPolyline)
-
-    entity = NDTool.build(entity_inference)
-    assert isinstance(entity, NDTextEntity)
-
-    mask = NDTool.build(segmentation_inference)
-    assert isinstance(mask, NDMask)
+@pytest_parametrize_plus("inference, expected_type",
+                         [(fixture_ref('polygon_inference'), NDPolygon),
+                          (fixture_ref('rectangle_inference'), NDRectangle),
+                          (fixture_ref('line_inference'), NDPolyline),
+                          (fixture_ref('entity_inference'), NDTextEntity),
+                          (fixture_ref('segmentation_inference'), NDMask)])
+def test_tool_construction(inference, expected_type):
+    assert isinstance(NDTool.build(inference), expected_type)
 
 
 def test_incorrect_feature_schema(rectangle_inference, polygon_inference,
@@ -47,7 +39,7 @@ def test_incorrect_feature_schema(rectangle_inference, polygon_inference,
     #Prob the error message says something about the config not anything useful. We might want to fix this.
     pred = rectangle_inference.copy()
     pred['schemaId'] = polygon_inference['schemaId']
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         _validate_ndjson([pred], configured_project)
 
 
@@ -55,14 +47,14 @@ def no_tool(text_inference, configured_project):
     pred = text_inference.copy()
     #Missing key
     del pred['answer']
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         _validate_ndjson([pred], configured_project)
 
 
 def test_invalid_datarow_id(text_inference, configured_project):
     pred = text_inference.copy()
     pred['dataRow']['id'] = "1232132131232"
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         _validate_ndjson([pred], configured_project)
 
 
@@ -72,18 +64,18 @@ def test_invalid_text(text_inference, configured_project):
     #Extra and wrong key
     del pred['answer']
     pred['answers'] = []
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         _validate_ndjson([pred], configured_project)
     del pred['answers']
 
     #Invalid type
     pred['answer'] = []
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         _validate_ndjson([pred], configured_project)
 
     #Invalid type
     pred['answer'] = None
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         _validate_ndjson([pred], configured_project)
 
 
@@ -92,23 +84,23 @@ def test_invalid_checklist_item(checklist_inference, configured_project):
     pred = checklist_inference.copy()
     pred['answers'] = [pred['answers'][0], pred['answers'][0]]
     #Duplicate schema ids
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         _validate_ndjson([pred], configured_project)
 
     pred['answers'] = [{"schemaId": "1232132132"}]
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         _validate_ndjson([pred], configured_project)
 
     pred['answers'] = [{}]
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         _validate_ndjson([pred], configured_project)
 
     pred['answers'] = []
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         _validate_ndjson([pred], configured_project)
 
     del pred['answers']
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         _validate_ndjson([pred], configured_project)
 
 
@@ -116,7 +108,7 @@ def test_invalid_polygon(polygon_inference, configured_project):
     #Only two points
     pred = polygon_inference.copy()
     pred['polygon'] = [{"x": 100, "y": 100}, {"x": 200, "y": 200}]
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         _validate_ndjson([pred], configured_project)
 
 
@@ -124,26 +116,26 @@ def test_incorrect_entity(entity_inference, configured_project):
     entity = entity_inference.copy()
     #Location cannot be a list
     entity["location"] = [0, 10]
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         _validate_ndjson([entity], configured_project)
 
     entity["location"] = {"start": -1, "end": 5}
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         _validate_ndjson([entity], configured_project)
 
     entity["location"] = {"start": 15, "end": 5}
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         _validate_ndjson([entity], configured_project)
 
 
 def test_incorrect_mask(segmentation_inference, configured_project):
     seg = segmentation_inference.copy()
     seg['mask']['colorRGB'] = [-1, 0, 10]
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         _validate_ndjson([seg], configured_project)
 
     seg['mask']['colorRGB'] = [0, 0]
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         _validate_ndjson([seg], configured_project)
 
 
@@ -156,13 +148,13 @@ def test_all_validate_json(configured_project, predictions):
 def test_incorrect_line(line_inference, configured_project):
     line = line_inference.copy()
     line["line"] = [line["line"][0]]  #Just one point
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         _validate_ndjson([line], configured_project)
 
 
 def test_incorrect_rectangle(rectangle_inference, configured_project):
     del rectangle_inference['bbox']['top']
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         _validate_ndjson([rectangle_inference], configured_project)
 
 
@@ -170,7 +162,7 @@ def test_duplicate_tools(rectangle_inference, configured_project):
     #Trying to upload a polygon and rectangle at the same time
     pred = rectangle_inference.copy()
     pred['polygon'] = [{"x": 100, "y": 100}, {"x": 200, "y": 200}]
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         _validate_ndjson([pred], configured_project)
 
 
@@ -178,7 +170,7 @@ def test_invalid_feature_schema(configured_project, rectangle_inference):
     #Trying to upload a polygon and rectangle at the same time
     pred = rectangle_inference.copy()
     pred['schemaId'] = "blahblah"
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         _validate_ndjson([pred], configured_project)
 
 
@@ -186,7 +178,7 @@ def test_missing_feature_schema(configured_project, rectangle_inference):
     #Trying to upload a polygon and rectangle at the same time
     pred = rectangle_inference.copy()
     del pred['schemaId']
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         _validate_ndjson([pred], configured_project)
 
 
@@ -212,11 +204,11 @@ def test_validate_ndjson_uuid(tmp_path, configured_project, predictions):
     with file_path.open("w") as f:
         ndjson.dump(repeat_uuid, f)
 
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         configured_project.upload_annotations(name="name",
                                               annotations=str(file_path))
 
-    with pytest.raises(NDJsonError):
+    with pytest.raises(ValidationError):
         configured_project.upload_annotations(name="name",
                                               annotations=repeat_uuid)
